Made depth a floating point variable

+++ au.c
@@ -159,12 +159,12 @@ void au_draw_texture_rect(AU_Engine* eng, AU_TextureRegion tex, AU_Rectangle rec
 }
 
 void au_draw_texture_transform(AU_Engine* eng, AU_TextureRegion tex, AU_Transform trans, float x, float y, float w,
-							   float h, int depth) {
+							   float h, float depth) {
 	au_draw_texture_blend(eng, tex, AU_WHITE, trans, x, y, w, h, false, false, depth);
 }
 
 void au_draw_texture_ex(AU_Engine* eng, AU_TextureRegion tex, AU_Color color, float x, float y, float w, float h,
-						float rot, float or_x, float or_y, float scale_x, float scale_y, bool flip_x, bool flip_y, int depth) {
+						float rot, float or_x, float or_y, float scale_x, float scale_y, bool flip_x, bool flip_y, float depth) {
 	AU_Transform trans = au_geom_identity();
 	trans = au_geom_transform_concat(trans, au_geom_transform_rotate(rot));
 	trans = au_geom_transform_concat(trans, au_geom_transform_scale(scale_x, scale_y));
@@ -173,7 +173,7 @@ void au_draw_texture_ex(AU_Engine* eng, AU_TextureRegion tex, AU_Color color, fl
 }
 
 void au_draw_texture_blend(AU_Engine* eng, AU_TextureRegion tex, AU_Color color, AU_Transform trans, float x, float y,
-						   float w, float h, bool flip_x, bool flip_y, int depth) {
+						   float w, float h, bool flip_x, bool flip_y, float depth) {
 	AU_Context* ctx = &(eng->ctx);
 
 	//Calculate the destination points with the transformation
